feat(datalog): update queue datatype to uint32

+++ datalog.c
@@ -9,7 +9,7 @@
 #include "datalog.h"
 
 // Declare the queue array and front, rear variables
-int queue[MAX_SIZE];
+uint32_t queue[MAX_SIZE];
 int front = -1, rear = -1;
 
 // Function to check if the queue is full
@@ -23,7 +23,7 @@ int isEmpty(){
 }
 
 // Function to enqueue (insert) an element
-void enqueue(unsigned long data){
+void enqueue(uint32_t data){
     // if queue full, do not enqueue
     if (isFull()) {
         return;
@@ -38,14 +38,14 @@ void enqueue(unsigned long data){
 }
 
 // Function to dequeue (remove) an element
-unsigned long dequeue()
+uint32_t dequeue()
 {
     // If the queue is empty, do not attempt to dequeue
     if (isEmpty()) {
         return 0;
     }
     // Get the data from the front of the queue
-    int data = queue[front];
+    uint32_t data = queue[front];
 
     // If the front and rear pointers are at the same
     // position, reset them
