+++ datalog.h
@@ -9,13 +9,15 @@
 #ifndef DATALOG_H
 #define DATALOG_H
 
+#include <stdint.h>
+
 // --- CONSTANTS ---
 #define MAX_SIZE 100        // Max array size of queue
 
 // Function Prototypes
 int isFull();
 int isEmpty();
-void enqueue(unsigned long data);
-unsigned long dequeue();
+void enqueue(uint32_t data);
+uint32_t dequeue();
 
 #endif /* DATALOG_H */
